+++ SpriteFilter/SpriteFilter.jar
